DockerFRTriton/pipeline.py
"""
Face Recognition Pipeline for Triton Inference Server.

Implementation Note:
This uses a pragmatic hybrid approach to balance Triton usage with implementation complexity:

- Face detection/alignment: CPU (InsightFace)
  Reason: RetinaFace ONNX output requires complex postprocessing (anchor decoding, NMS)
  that is beyond the scope of this assignment. The detection model IS deployed to Triton
  (see model_repository/face_detector) but full integration requires significant work.

- Embedding extraction: Triton (ONNX model) ✓
  This is the PRIMARY inference task and runs fully on Triton as required.

Alternative: For a pure Triton solution, students would need to implement:
1. Anchor generation for RetinaFace
2. Non-maximum suppression (NMS)
3. Keypoint postprocessing
All of which are non-trivial and beyond the FR model focus of this assignment.

The current approach fulfills the core requirement: FR model inference via Triton.
"""

from typing import Any, Tuple

import cv2
import numpy as np

from triton_service import run_embedding

# ArcFace alignment template (from DockerFR/util.py)
ARCFACE_DST = np.array([
    [38.2946, 51.6963],  # Left eye
    [73.5318, 51.5014],  # Right eye
    [56.0252, 71.7366],  # Nose
    [41.5493, 92.3655],  # Left mouth
    [70.7299, 92.2041]   # Right mouth
], dtype=np.float32)


def _cosine_similarity(vec_a: np.ndarray, vec_b: np.ndarray) -> float:
    """Compute cosine similarity between two 1D vectors."""
    a_norm = np.linalg.norm(vec_a)
    b_norm = np.linalg.norm(vec_b)
    if a_norm == 0.0 or b_norm == 0.0:
        return 0.0
    return float(np.dot(vec_a, vec_b) / (a_norm * b_norm))


def _bytes_to_image(image_bytes: bytes) -> np.ndarray:
    """Convert image bytes to BGR numpy array."""
    from PIL import Image
    import io

    try:
        image = Image.open(io.BytesIO(image_bytes))
        image_array = np.array(image)

        if len(image_array.shape) == 3 and image_array.shape[2] == 3:
            image_bgr = cv2.cvtColor(image_array, cv2.COLOR_RGB2BGR)
        elif len(image_array.shape) == 2:
            image_bgr = cv2.cvtColor(image_array, cv2.COLOR_GRAY2BGR)
        else:
            image_bgr = cv2.cvtColor(image_array, cv2.COLOR_RGBA2BGR)
        return image_bgr
    except Exception:
        nparr = np.frombuffer(image_bytes, np.uint8)
        img = cv2.imdecode(nparr, cv2.IMREAD_COLOR)
        if img is None:
            raise ValueError("Failed to decode image")
        return img


def _detect_face_with_insightface(image: np.ndarray):
    """
    Use InsightFace (CPU) for face detection and keypoint extraction.
    Returns best face keypoints or raises error if no face found.
    """
    try:
        from insightface.app import FaceAnalysis
    except ImportError:
        raise RuntimeError("InsightFace required. Install: pip install insightface")

    # Initialize InsightFace
    app = FaceAnalysis(name='buffalo_l', providers=['CPUExecutionProvider'])
    app.prepare(ctx_id=-1, det_size=(640, 640))

    # Detect faces
    faces = app.get(image)
    if not faces:
        raise ValueError("No face detected in image")

    # Get best face
    best_face = max(faces, key=lambda f: f.det_score)
    return best_face.kps  # 5x2 landmarks


def _align_face(image: np.ndarray, keypoints: np.ndarray) -> np.ndarray:
    """Align face to 112x112 using 5-point landmarks."""
    transformation_matrix, _ = cv2.estimateAffinePartial2D(
        keypoints, ARCFACE_DST, method=cv2.LMEDS
    )

    if transformation_matrix is None:
        raise ValueError("Failed to compute transformation matrix")

    aligned_face = cv2.warpAffine(
        image, transformation_matrix, (112, 112),
        flags=cv2.INTER_LINEAR, borderMode=cv2.BORDER_CONSTANT, borderValue=0
    )
    return aligned_face


def _antispoof_check(aligned_face: np.ndarray) -> Tuple[bool, float]:
    """
    Simple anti-spoofing check using image quality heuristics.

    Returns:
        (is_real, confidence_score) where:
        - is_real: True if face appears to be real (live)
        - confidence_score: 0.0-1.0, higher means more confident it's real

    Note: This is a basic implementation. For production, consider:
    - Deep learning models (e.g., FaceX-Zoo, Silent-Face-Anti-Spoofing)
    - Multi-modal checks (texture, motion, depth)
    - Dedicated anti-spoofing models on Triton
    """
    # Check 1: Image sharpness (Laplacian variance)
    gray = cv2.cvtColor(aligned_face, cv2.COLOR_BGR2GRAY)
    laplacian_var = cv2.Laplacian(gray, cv2.CV_64F).var()
    sharpness_score = min(laplacian_var / 500.0, 1.0)  # Normalize

    # Check 2: Color diversity (std of color channels)
    color_std = np.std(aligned_face, axis=(0, 1)).mean()
    color_score = min(color_std / 50.0, 1.0)  # Normalize

    # Check 3: Brightness check (avoid overexposed/underexposed)
    mean_brightness = gray.mean()
    brightness_score = 1.0 - abs(mean_brightness - 127.5) / 127.5

    # Combine scores (weighted average)
    confidence = (sharpness_score * 0.5 + color_score * 0.3 + brightness_score * 0.2)

    # Threshold for real vs spoof
    is_real = confidence > 0.4  # Adjust threshold as needed

    return is_real, confidence


def get_embeddings(client: Any, image_a: bytes, image_b: bytes) -> Tuple[np.ndarray, np.ndarray]:
    """
    Extract embeddings for two images using hybrid approach:
    - Detection/alignment: CPU (InsightFace)
    - Anti-spoofing: CPU (image quality heuristics)
    - Embedding: Triton (ONNX)
    """
    # Process image A
    img_a = _bytes_to_image(image_a)
    keypoints_a = _detect_face_with_insightface(img_a)
    aligned_a = _align_face(img_a, keypoints_a)

    # Anti-spoofing check for image A
    is_real_a, confidence_a = _antispoof_check(aligned_a)
    if not is_real_a:
        raise ValueError(f"Image A failed anti-spoofing check (confidence: {confidence_a:.2f})")

    emb_a = run_embedding(client, aligned_a)  # Triton

    # Process image B
    img_b = _bytes_to_image(image_b)
    keypoints_b = _detect_face_with_insightface(img_b)
    aligned_b = _align_face(img_b, keypoints_b)

    # Anti-spoofing check for image B
    is_real_b, confidence_b = _antispoof_check(aligned_b)
    if not is_real_b:
        raise ValueError(f"Image B failed anti-spoofing check (confidence: {confidence_b:.2f})")

    emb_b = run_embedding(client, aligned_b)  # Triton

    return emb_a, emb_b


def calculate_face_similarity(client: Any, image_a: bytes, image_b: bytes) -> float:
    """
    End-to-end face similarity pipeline:
    1. Detect faces + keypoints (CPU - InsightFace)
    2. Align to 112x112 (CPU - OpenCV)
    3. Anti-spoofing check (CPU - image quality heuristics)
    4. Extract embeddings (Triton - ONNX)
    5. Compute similarity (CPU)
    """
    emb_a, emb_b = get_embeddings(client, image_a, image_b)
    return _cosine_similarity(emb_a, emb_b)
